Refactor the APL UI.
Redesigned the APL UI to show the current accent being rendered
and a list of different accent the user can try.

This is not complete yet. For ex, we always shows the Japanese
image for the current accent. The text in the list of accents the
user can try also needs some tweaking. However, the basic framework
is all done and hence committing this change.